// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const OrganizationType = {
  HEADQUARTERS: 'HEADQUARTERS',
  MINISTRY: 'MINISTRY',
  REGION: 'REGION',
  ZONE: 'ZONE',
  WOREDA: 'WOREDA',
  KEBELE: 'KEBELE',
  DEPARTMENT: 'DEPARTMENT',
  BRANCH: 'BRANCH',
  PROJECT: 'PROJECT',
  EXTERNAL_PARTNER: 'EXTERNAL_PARTNER'
} as const

export type OrganizationType = (typeof OrganizationType)[keyof typeof OrganizationType]


export const AccountStatus = {
  PENDING: 'PENDING',
  ACTIVE: 'ACTIVE',
  SUSPENDED: 'SUSPENDED',
  LOCKED: 'LOCKED',
  ARCHIVED: 'ARCHIVED'
} as const

export type AccountStatus = (typeof AccountStatus)[keyof typeof AccountStatus]


export const UserType = {
  INTERNAL: 'INTERNAL',
  EXTERNAL: 'EXTERNAL',
  CONTRACTOR: 'CONTRACTOR',
  CITIZEN: 'CITIZEN',
  API: 'API',
  SYSTEM: 'SYSTEM'
} as const

export type UserType = (typeof UserType)[keyof typeof UserType]


export const PermissionEffect = {
  ALLOW: 'ALLOW',
  DENY: 'DENY'
} as const

export type PermissionEffect = (typeof PermissionEffect)[keyof typeof PermissionEffect]


export const ScopeType = {
  GLOBAL: 'GLOBAL',
  TENANT: 'TENANT',
  ORGANIZATION: 'ORGANIZATION',
  DESCENDANT: 'DESCENDANT',
  DEPARTMENT: 'DEPARTMENT',
  SELF: 'SELF'
} as const

export type ScopeType = (typeof ScopeType)[keyof typeof ScopeType]


export const MfaMethodType = {
  TOTP: 'TOTP',
  EMAIL: 'EMAIL',
  SMS: 'SMS',
  WEBAUTHN: 'WEBAUTHN'
} as const

export type MfaMethodType = (typeof MfaMethodType)[keyof typeof MfaMethodType]


export const AuditActionType = {
  CREATE: 'CREATE',
  UPDATE: 'UPDATE',
  DELETE: 'DELETE',
  LOGIN: 'LOGIN',
  LOGOUT: 'LOGOUT',
  PERMISSION_CHECK: 'PERMISSION_CHECK',
  ROLE_ASSIGNED: 'ROLE_ASSIGNED',
  POLICY_APPLIED: 'POLICY_APPLIED'
} as const

export type AuditActionType = (typeof AuditActionType)[keyof typeof AuditActionType]